pyproject.toml
[tool.poetry]
name = "langchain-crash-course"
version = "0.1.0"
description = "Everything you need to know to get started with LangChain"
authors = ["bhancock_ai <brandon@brandonhancock.io>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12.5"
# langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
# langchain-anthropic = "^0.1.15"
# langchain-google-genai = "^1.0.5"
# langchain-google-firestore = "^0.3.0"
# firestore = "^0.0.8"
# tiktoken = "^0.7.0"
# firecrawl-py = "^0.0.13"
# langchainhub = "^0.1.18"
# wikipedia = "^1.4.0"
# tavily-python = "^0.3.3"
fastembed-gpu = "^0.3.1"
chromadb = "^0.5.3"
protobuf = "==3.20.*"
onnxruntime-gpu = "==1.18.1"
sentence-transformers = "^3.0.1"
tiktoken = "^0.7.0"
langchain-huggingface = "^0.0.3"
torch = "^2.3.1"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.20"
wikipedia = "^1.4.0"
langchainhub = "^0.1.20"
langchain-openai = "^0.1.14"

[[tool.poetry.source]]
name = "onnxruntime-gpu"
url = "https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/"
priority = "supplemental"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 
